+++ docs/changelog.md
@@ -22,6 +22,7 @@ Sprint 9 / March 28, 2020
 - added label to records.
 - added tests for Dashboard page.
 - added React.FC typing to all functional components.
+- updated action name button clicked to stop in dashboard component.
 
 ## Release 0.8.0
 Sprint 8 / March 14, 2020